RDM-4879 cleanup code
