+++ config/person_key_server_config.py
@@ -1,0 +1,25 @@
+VERSION = '0.0.1'
+EXCLUDED_FILES = [".*", 
+                  "Thumbs.db", 
+                  ".DS_Store", 
+                  "._.DS_Store", 
+                  ".localized", 
+                  "desktop.ini", 
+                  "*.pyc", 
+                  "*.swx",
+                  "*.swp", 
+                  "*~", 
+                  "~$*", 
+                  "NULLEXT"]
+APP_NAME = "fdrd-encryption-client"
+APP_AUTHOR = "Simon Fraser University"
+TIMEZONE = "America/Regina"
+VAULT_DATASET_KEY_PATH = "dataset"
+VAULT_PUBLIC_KEY_PATH = "public_key"
+VAULT_GENERATE_CERT_ROLE_NAME = "example-dot-com"
+VAULT_GENERATE_CERT_COMMON_NAME = "test.206-12-90-40.cloud.computecanada.ca"
+VAULT_GENERATE_CERT_MOUNT_NAME = "pki_int"
+LOCAL_KEY_DIR_NAME = "frdr_keys"
+LOCAL_PRIVATE_KEY_FILENAME = "private.key"
+LOCAL_PUBLIC_KEY_FILENAME = "public.key"
+
