+++ modules/PersonKeyManager.py
@@ -1,9 +1,7 @@
 import os
-import nacl
-import hvac
 import logging
 from util.util import Util
-from util import constants
+from config import app_config
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives.asymmetric import rsa
 from cryptography.hazmat.primitives import serialization
@@ -11,11 +9,11 @@ from cryptography.hazmat.primitives import hashes
 from cryptography.hazmat.primitives.asymmetric import padding
 from cryptography import x509
 
-Util.get_logger("frdr-crypto.public-key-manager")
+Util.get_logger("fdrd-encryption-client.person-key-manager")
 
-class PublicKeyManager(object):
+class PersonKeyManager(object):
     def __init__(self, vault_client):
-        self._logger = logging.getLogger("frdr-crypto.public-key-mamanger.vault")
+        self._logger = logging.getLogger("fdrd-encryption-client.person-key-manager.vault")
         self._vault_client = vault_client
         self._public_key = None
 
@@ -27,9 +25,9 @@ class PublicKeyManager(object):
 
     def generate_certificate(self):
         certificate, private_key_str = self._vault_client.generate_certificate(
-            name=constants.VAULT_GENERATE_CERT_ROLE_NAME,
-            common_name=constants.VAULT_GENERATE_CERT_COMMON_NAME,
-            mount_point=constants.VAULT_GENERATE_CERT_MOUNT_NAME
+            name=app_config.VAULT_GENERATE_CERT_ROLE_NAME,
+            common_name=app_config.VAULT_GENERATE_CERT_COMMON_NAME,
+            mount_point=app_config.VAULT_GENERATE_CERT_MOUNT_NAME
         )
         private_key = private_key_str.encode()
         return (certificate, private_key)
@@ -57,14 +55,14 @@ class PublicKeyManager(object):
         return public_key
 
     def save_public_key_to_vault(self, public_key):
-        path = "/".join([constants.VAULT_PUBLIC_KEY_PATH, self.get_vault_entity_id()])
+        path = "/".join([app_config.VAULT_PUBLIC_KEY_PATH, self.get_vault_entity_id()])
         if isinstance(public_key, bytes):
             public_key = Util.byte_to_base64(public_key)
         self._vault_client.save_key_to_vault(path, public_key)
 
     def read_public_key_from_vault(self, user_entity_id):
         try:
-            path = "/".join([constants.VAULT_PUBLIC_KEY_PATH, user_entity_id])
+            path = "/".join([app_config.VAULT_PUBLIC_KEY_PATH, user_entity_id])
             return self._vault_client.retrive_key_from_vault(path)
         except Exception as e:
             self._logger.error("Falied to read public key from Vault. {}".format(e))
@@ -78,10 +76,10 @@ class PublicKeyManager(object):
         if public_key_on_vault is None:
             self._logger.info("No public key is saved on Vault for the current user, generating a pair of public and private key")
             cert, private_key = self.generate_certificate()
-            private_key_path = os.path.join(Util.get_key_dir(self.get_vault_entity_id()), constants.LOCAL_PRIVATE_KEY_FILENAME)
+            private_key_path = os.path.join(Util.get_key_dir(self.get_vault_entity_id()), app_config.LOCAL_PRIVATE_KEY_FILENAME)
             self.save_key_locally(private_key, private_key_path)
             public_key = self.extract_public_key_from_cert(cert)
-            public_key_path = os.path.join(Util.get_key_dir(self.get_vault_entity_id()), constants.LOCAL_PUBLIC_KEY_FILENAME)
+            public_key_path = os.path.join(Util.get_key_dir(self.get_vault_entity_id()), app_config.LOCAL_PUBLIC_KEY_FILENAME)
             self.save_key_locally(public_key, public_key_path)
             self.save_public_key_to_vault(public_key)
             return public_key
